rss_link: serviceにload,convertの統合
流石にファイルが増え過ぎた
